+++ app/internal/dependencies.py
@@ -1,0 +1,98 @@
+# hashing and jwt
+from jose import JWTError, jwt
+from datetime import timedelta, datetime
+from passlib.context import CryptContext
+
+# fastapi
+from fastapi import Depends, HTTPException, status
+from fastapi.security import OAuth2PasswordBearer
+
+# database
+from pymongo import MongoClient
+from internal.validateDBConnection import validateMongo
+from internal.connectionString import CONNECTION_STRING
+
+# models
+from models import UserInDB, User, TokenData
+
+# environement variables for toekn creation
+from internal.env import SECRET_KEY, ALGORITHM #NEEDS TO BE CHANGED FOR PROD *********
+
+
+pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
+
+oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")
+
+
+def verify_password(plain_password, hashed_password):
+    return pwd_context.verify(plain_password, hashed_password)
+
+
+def get_password_hash(password):
+    return pwd_context.hash(password)
+
+
+# This is the method that accesses the user database 
+def get_user(username: str):    
+    client = MongoClient(CONNECTION_STRING)
+    validateMongo(client)
+    user_dict = client.api.Users.find_one({'username': username}) #database = api, collection = Users
+    return UserInDB(**user_dict) #returns None if not found
+    
+
+def authenticate_user(username: str, password: str):
+    user = get_user(username)
+    if not user:
+        return False
+    if not verify_password(password, user.hashed_password):
+        return False
+    return user
+
+
+def create_access_token(data: dict, expires_delta: timedelta | None = None):
+    to_encode = data.copy()
+    if expires_delta:
+        expire = datetime.utcnow() + expires_delta
+    else:
+        expire = datetime.utcnow() + timedelta(minutes=15)
+    to_encode.update({"exp": expire})
+    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
+    return encoded_jwt
+
+
+async def get_current_user(token: str = Depends(oauth2_scheme)):
+    credentials_exception = HTTPException(
+        status_code=status.HTTP_401_UNAUTHORIZED,
+        detail="Could not validate credentials",
+        headers={"WWW-Authenticate": "Bearer"},
+    )
+    try:
+        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
+        username: str = payload.get("sub")
+        if username is None:
+            raise credentials_exception
+        token_data = TokenData(username=username)
+    except JWTError:
+        raise credentials_exception
+    user = get_user(username=token_data.username)
+    if user is None:
+        raise credentials_exception
+    return user
+
+
+async def get_current_active_user(current_user: User = Depends(get_current_user)):
+    if current_user.disabled:
+        raise HTTPException(status_code=400, detail="Inactive user")
+    return current_user
+
+
+
+async def get_current_admin_user(current_user: User = Depends(get_current_active_user)):
+    credentials_exception = HTTPException(
+        status_code=status.HTTP_401_UNAUTHORIZED,
+        detail="User does not have administrative access",
+        headers={"WWW-Authenticate": "Bearer"},
+    )
+    if not current_user.admin:
+        raise credentials_exception
+    return current_user
