+++ app/models.py
@@ -1,35 +1,50 @@
 from pydantic import BaseModel
 
+# token models
+class Token(BaseModel):
+    access_token: str
+    token_type: str
+
+
+class TokenData(BaseModel):
+    username: str | None = None
+
+
+# user models
 class User(BaseModel):
     _id: int | None = None
     username: str
     email: str | None = None
     full_name: str | None = None
     disabled: bool | None = None
+    admin: bool | None = None 
 
 
-# additional pw fields for database user model
 class UserInDB(User):
-    hashed_password: str
+    hashed_password: str # additional pw fields for database user model
     
 
+# items model
 class Items(BaseModel):
     items: list[dict]
     count: int
 
 
+# version information model
 class Version(BaseModel):
     version: str
     author: str
     repository: str
 
 
+# oil model
 class Oil(BaseModel):
     _id: int | None = None
     name: str 
     sapratio: float
 
 
+# Recipes model
 class Recipe(BaseModel):
     _id: int | None = None
     name: str
